feat: Use u64 for prelude collection indices
Return u64 from length and accept u64 counts and indices in take, skip,
and get. Preserve full unsigned values in bounds errors and update the
specification and regression coverage.

+++ rex-engine/src/error.rs
@@ -190,7 +190,7 @@ pub enum EngineError {
     #[error("index {index} out of bounds in `{name}` (len {len})")]
     IndexOutOfBounds {
         name: Symbol,
-        index: i32,
+        index: i128,
         len: usize,
     },
     #[error("internal error: {0}")]
